Move configurator from homepage to Kairos Factory page
And display the config next to the kairos-lab installation

Fixes kairos-io/kairos#4040
