+++ docs/reference/kairos-factory.mdx
@@ -3,6 +3,7 @@ title: "The Kairos Factory"
 sidebar_label: "The Kairos Factory"
 sidebar_position: 4
 description: "Converting any base image into a Kairos ready image"
+show_image_builder: true
 ---
 
 :::warning Ongoing Project
